resolving bugs

+++ emcare-ui/src/app/auth/auth.guard.ts
@@ -10,17 +10,25 @@ export class AuthGuard implements CanActivate {
     ) { }
 
     canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean | UrlTree | Observable<boolean | UrlTree> | Promise<boolean | UrlTree> {
+        const token = JSON.parse(localStorage.getItem('access_token'));
         const tokenExpiryDate = JSON.parse(localStorage.getItem('refresh_token_expiry_time'));
         const tokenExpiry = tokenExpiryDate
             ? new Date(tokenExpiryDate)
             : null;
         // Check if token is expired or not
-        if (tokenExpiry && tokenExpiry <= new Date()) {
-            // token has expired user should be logged out
-            this.router.navigate(['/login']);
-            localStorage.clear();
-            return false;
+        if (token) {
+            if (tokenExpiry && tokenExpiry <= new Date()) {
+                // token has expired user should benavigateByUrl logged out
+                this.router.navigate(['/login']);
+                localStorage.clear();
+                return false;
+            } else {
+                return true;
+            }
         } else {
+            if (route.routeConfig.path !== 'login') {
+                this.router.navigate(['/login']);
+            }
             return true;
         }
     }
